+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.1.0 2021-06-26
+
+### Added
+- exported `PouchDB` from `rxdb` to our plugin.
+- exported `(find, findOne, database)` to our plugin.
+- added new plugin `pouchdb-adapter-http` for sync.
+- added vuex actions from module.
+
+### Changed
+- changed search regex query.
+
 ## 2.0.4 2021-05-31
 ### Fixed
 - Fix RxJS warn